Turn keess annotations check into a reusable matcher
