+++ tests/configmap_sync_test.go
@@ -4,9 +4,11 @@ import (
 	"context"
 	"fmt"
 	"path/filepath"
+	"reflect"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/types"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -46,7 +48,10 @@ var _ = Describe("ConfigMap Sync", func() {
 				By("Waiting for ConfigMap to be synchronized")
 				Eventually(getConfigMap, syncTimeout, pollInterval).WithArguments(
 					destinationClusterClient, configMapName, configMapNamespace).Should(
-					BeEqualToSourceConfigMap(), fmt.Sprintf("ConfigMap %s/%s should exist within %v and match source configmap", configMapNamespace, configMapName, syncTimeout))
+					And(
+						BeEqualToSourceConfigMap(),
+						WithTransform(getConfigMapMeta, HaveKeessTrackingAnnotations(configMapNamespace)),
+					), fmt.Sprintf("ConfigMap %s/%s should exist within %v and match source configmap", configMapNamespace, configMapName, syncTimeout))
 			})
 		})
 
@@ -58,8 +63,8 @@ var _ = Describe("ConfigMap Sync", func() {
 
 				By("Waiting for ConfigMap to be synchronized")
 				Eventually(getConfigMap, syncTimeout, pollInterval).WithArguments(
-					destinationClusterClient, configMapName, configMapNamespace).Should(
-					BeEqualToSourceConfigMap(), fmt.Sprintf("ConfigMap %s/%s should exist within %v and match source configmap", configMapNamespace, configMapName, syncTimeout))
+					destinationClusterClient, configMapName, configMapNamespace).Should(Not(BeNil()),
+					fmt.Sprintf("ConfigMap %s/%s should exist within %v and match source configmap", configMapNamespace, configMapName, syncTimeout))
 
 				By("Updating ConfigMap in source cluster")
 				// we know there is no error because of the previous Eventually check
@@ -80,6 +85,7 @@ var _ = Describe("ConfigMap Sync", func() {
 					destinationClusterClient, configMapName, configMapNamespace).Should(
 					And(
 						BeEqualToSourceConfigMap(),
+						WithTransform(getConfigMapMeta, HaveKeessTrackingAnnotations(configMapNamespace)),
 						WithTransform(getConfigMapData, HaveKeyWithValue("logging.level", "DEBUG")),
 						WithTransform(getConfigMapData, HaveKeyWithValue("new.key", "new.value")),
 					), fmt.Sprintf("ConfigMap %s/%s should be updated within %v", configMapNamespace, configMapName, syncTimeout))
@@ -90,3 +96,21 @@ var _ = Describe("ConfigMap Sync", func() {
 	})
 	//TODO: Context("On Namespace mode", func() {})
 })
+
+// Get metadata from Secret object
+func getConfigMapMeta(config *corev1.ConfigMap) *metav1.ObjectMeta {
+	return &config.ObjectMeta
+}
+
+// Custom matcher to check if a ConfigMap on destination cluster matches the source ConfigMap
+func BeEqualToSourceConfigMap() types.GomegaMatcher {
+	return WithTransform(func(configmap *corev1.ConfigMap) bool {
+		sourceConfigMap, err := sourceClusterClient.CoreV1().ConfigMaps(configmap.Namespace).Get(context.Background(), configmap.Name, metav1.GetOptions{})
+		if err != nil {
+			return false
+		}
+
+		// Compare only the Data field, ignoring metadata differences
+		return reflect.DeepEqual(configmap.Data, sourceConfigMap.Data)
+	}, BeTrue())
+}
