+++ tests/e2e_suite_test.go
@@ -149,53 +149,12 @@ func deleteNamespaceOnAll(namespace string, wait bool) {
 	deleteNamespace(destinationClusterClient, namespace, wait)
 }
 
-// Custom matcher to check if a Secret on destination cluster matches the source Secret
-func BeEqualToSourceSecret() types.GomegaMatcher {
-	return WithTransform(func(secret *corev1.Secret) bool {
-		sourceSecret, err := sourceClusterClient.CoreV1().Secrets(secret.Namespace).Get(context.Background(), secret.Name, metav1.GetOptions{})
-		if err != nil {
-			return false
-		}
-
-		// Secret Sync actually DOES NOT sync labels and annotations. Not sure if that's intended.
-		// TODO: is it a bug?
-
-		// // Check that all labels from source are present in the destination Secret
-		// for key, value := range sourceSecret.Labels {
-		// 	Expect(secret.Labels).To(HaveKeyWithValue(key, value), fmt.Sprintf("Label %s should match source Secret", key))
-		// }
-
-		// // Check that all annotations from source are present in the destination Secret
-		// for key, value := range sourceSecret.Annotations {
-		// 	Expect(secret.Annotations).To(HaveKeyWithValue(key, value), fmt.Sprintf("Annotation %s should match source Secret", key))
-		// }
-
-		Expect(secret.Labels).To(HaveKeyWithValue("keess.powerhrg.com/managed", "true"), "Destination Secret should have correct managed label")
-		Expect(secret.Annotations).To(HaveKeyWithValue("keess.powerhrg.com/source-cluster", sourceClusterContext), "Destination Secret should have correct source cluster annotation")
-		Expect(secret.Annotations).To(HaveKeyWithValue("keess.powerhrg.com/source-namespace", sourceSecret.Namespace), "Destination Secret should have correct source namespace annotation")
-
-		// TODO: I think we found a bug here, because the source resource version is not synced whe source is updated
-		// This line catches that when on the update case
-		// Expect(secret.Annotations).To(HaveKeyWithValue("keess.powerhrg.com/source-resource-version", sourceSecret.ResourceVersion), "Destination Secret should have correct source resource version annotation")
-
-		// Compare only the Data field, ignoring metadata differences
-		return reflect.DeepEqual(secret.Data, sourceSecret.Data)
-	}, BeTrue())
-}
-
-// Custom matcher to check if a ConfigMap on destination cluster matches the source ConfigMap
-func BeEqualToSourceConfigMap() types.GomegaMatcher {
-	return WithTransform(func(configmap *corev1.ConfigMap) bool {
-		sourceConfigMap, err := sourceClusterClient.CoreV1().ConfigMaps(configmap.Namespace).Get(context.Background(), configmap.Name, metav1.GetOptions{})
-		if err != nil {
-			return false
-		}
-
-		Expect(configmap.Labels).To(HaveKeyWithValue("keess.powerhrg.com/managed", "true"), "Destination ConfigMap should have correct managed label")
-		Expect(configmap.Annotations).To(HaveKeyWithValue("keess.powerhrg.com/source-cluster", sourceClusterContext), "Destination ConfigMap should have correct source cluster annotation")
-		Expect(configmap.Annotations).To(HaveKeyWithValue("keess.powerhrg.com/source-namespace", sourceConfigMap.Namespace), "Destination ConfigMap should have correct source namespace annotation")
-
-		// Compare only the Data field, ignoring metadata differences
-		return reflect.DeepEqual(configmap.Data, sourceConfigMap.Data)
+// Custom matcher to check if metadata has the Keess tracking annotations
+func HaveKeessTrackingAnnotations(sourceNamespace string) types.GomegaMatcher {
+	return WithTransform(func(metadata *metav1.ObjectMeta) bool {
+		Expect(metadata.Labels).To(HaveKeyWithValue("keess.powerhrg.com/managed", "true"), "Destination object should have correct managed label")
+		Expect(metadata.Annotations).To(HaveKeyWithValue("keess.powerhrg.com/source-cluster", sourceClusterContext), "Destination object should have correct source cluster annotation")
+		Expect(metadata.Annotations).To(HaveKeyWithValue("keess.powerhrg.com/source-namespace", sourceNamespace), "Destination object should have correct source namespace annotation")
+		return true
 	}, BeTrue())
 }
